+++ src/aks-preview/HISTORY.md
@@ -6,6 +6,12 @@ Release History
 ++++++
 * Add support for `--format` parameter in `az aks get-credentials` command.  
 
+0.5.61
+++++++
+
+* Add support to create cluster with managed cluster snapshot. Command is
+  * `az aks create --cluster-snapshot-id <snapshot-id>`
+
 0.5.60
 ++++++
 
@@ -236,7 +242,7 @@ and `az aks nodepool add`
 0.5.24
 +++++
 
-* * Add "--aks-custom-headers" for "az aks nodepool upgrade"
+  * * Add "--aks-custom-headers" for "az aks nodepool upgrade"
 
 0.5.23
 +++++